HTMLQuoteElement quote delimiters should be based on lang of parent
Following the resolution of WHATWG HTML standard [1] and
existing CSS issue [2], it has been decided that the
auto (default) lang for <q> element should be based on the parent
language, and not the language of the element itself.

This CL makes sure to skip reading the lang attribute for <q>.
Instead, its locale is calculated in
CollectExtraStyleForPresentationAttribute(), using the parent's locale.

[1] https://github.com/whatwg/html/issues/3636
[2] https://github.com/w3c/csswg-drafts/issues/5478

Change-Id: Ief17dd1e7f5a40112899ce898fa672a6fef6e834
Fixed: 1290851
Bug: 753671